WIP Snake Browser Game
This is what I've been working on the past few days... I thought it would be a bit smaller than it is. Currently has no gameplay at all. But I just finished the beginings of the game's display with a button to test it. This also doesn't represent the two other attempts I made at building this. I had to change tact a few times. 

This is three days of effort with no instruction, the first project I came up with and wrote on my own in Javascript.